requirements/dev.txt
# pull in Common Packages
-r common.txt

coverage==4.0.3
django-debug-toolbar==1.4
factory-boy==2.6.1
fake-factory==0.5.7
flake8==2.5.4
funcsigs==1.0.0
ipaddress==1.0.16
mccabe==0.4.0
mock==2.0.0
pbr==1.8.1
pep8==1.7.0
pyflakes==1.0.0